requirements.txt
# ===== 프론트엔드 =====
streamlit>=1.31.0

# ===== LangChain 코어 =====
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# ===== LLM 제공자 =====
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.2.0

# ===== 임베딩 =====
# torch는 Dockerfile에서 CPU 버전으로 별도 설치 (용량 최적화)
langchain-huggingface>=0.1.0
sentence-transformers>=3.0.0

# ===== Vector DB (ChromaDB) =====
langchain-chroma>=0.1.0
chromadb>=0.5.0

# ===== 문서 파싱 =====
pypdf>=4.0.0                  # PDF 파싱 (순수 Python)
beautifulsoup4>=4.12.0        # 웹 HTML 파싱
requests>=2.31.0              # HTTP 요청
lxml>=5.0.0                   # HTML/XML 파서 (bs4 backend)

# ===== 유틸리티 =====
python-dotenv>=1.0.0          # .env 파일 로드
